+++ jsonschema/tests/test_validators.py
@@ -1,246 +1,80 @@
 from __future__ import unicode_literals
-from decimal import Decimal
 import contextlib
-import glob
-import io
 import json
-import os
-import re
-import subprocess
-import sys
+import pprint
+import textwrap
 
-if sys.version_info[:2] < (2, 7):  # pragma: no cover
-    import unittest2 as unittest
-else:
-    import unittest
-
-try:
-    from unittest import mock
-except ImportError:
-    import mock
-
-try:
-    from sys import pypy_version_info
-except ImportError:
-    pypy_version_info = None
-
-from jsonschema import (
-    PY3, SchemaError, UnknownType, ValidationError, ErrorTree,
-    Draft3Validator, Draft4Validator, FormatChecker, draft3_format_checker,
-    draft4_format_checker, RefResolver, validate, FormatError
+from jsonschema import FormatChecker, ValidationError
+from jsonschema.compat import PY3
+from jsonschema.tests.compat import mock, unittest
+from jsonschema.validators import (
+    RefResolutionError, UnknownType, ErrorTree, Draft3Validator,
+    Draft4Validator, RefResolver, create, extend, validator_for, validate,
 )
 
 
-THIS_DIR = os.path.dirname(__file__)
-TESTS_DIR = os.path.join(THIS_DIR, "json", "tests")
-
-JSONSCHEMA_SUITE = os.path.join(THIS_DIR, "json", "bin", "jsonschema_suite")
-
-REMOTES = subprocess.Popen(
-    ["python", JSONSCHEMA_SUITE, "remotes"], stdout=subprocess.PIPE,
-).stdout
-if PY3:
-    REMOTES = io.TextIOWrapper(REMOTES)
-REMOTES = json.load(REMOTES)
-
-
-def make_case(schema, data, valid):
-    if valid:
-        def test_case(self):
-            kwargs = getattr(self, "validator_kwargs", {})
-            validate(data, schema, cls=self.validator_class, **kwargs)
-    else:
-        def test_case(self):
-            kwargs = getattr(self, "validator_kwargs", {})
-            with self.assertRaises(ValidationError):
-                validate(data, schema, cls=self.validator_class, **kwargs)
-    return test_case
-
-
-def load_json_cases(tests_glob, ignore_glob="", basedir=TESTS_DIR, skip=None):
-    if ignore_glob:
-        ignore_glob = os.path.join(basedir, ignore_glob)
-
-    def add_test_methods(test_class):
-        ignored = set(glob.iglob(ignore_glob))
-
-        for filename in glob.iglob(os.path.join(basedir, tests_glob)):
-            if filename in ignored:
-                continue
-
-            validating, _ = os.path.splitext(os.path.basename(filename))
-
-            with open(filename) as test_file:
-                data = json.load(test_file)
-
-                for case in data:
-                    for test in case["tests"]:
-                        a_test = make_case(
-                            case["schema"],
-                            test["data"],
-                            test["valid"],
-                        )
-
-                        test_name = "test_%s_%s" % (
-                            validating,
-                            re.sub(r"[\W ]+", "_", test["description"]),
-                        )
-
-                        if not PY3:
-                            test_name = test_name.encode("utf-8")
-                        a_test.__name__ = test_name
-
-                        if skip is not None and skip(case):
-                            a_test = unittest.skip("Checker not present.")(
-                                a_test
-                            )
-
-                        setattr(test_class, test_name, a_test)
-
-        return test_class
-    return add_test_methods
-
-
-class TypesMixin(object):
-    @unittest.skipIf(PY3, "In Python 3 json.load always produces unicode")
-    def test_string_a_bytestring_is_a_string(self):
-        self.validator_class({"type" : "string"}).validate(b"foo")
-
-
-class DecimalMixin(object):
-    def test_it_can_validate_with_decimals(self):
-        schema = {"type" : "number"}
-        validator = self.validator_class(
-            schema, types={"number" : (int, float, Decimal)}
-        )
-
-        for valid in [1, 1.1, Decimal(1) / Decimal(8)]:
-            validator.validate(valid)
-
-        for invalid in ["foo", {}, [], True, None]:
-            with self.assertRaises(ValidationError):
-                validator.validate(invalid)
-
-
-def missing_format(checker):
-    def missing_format(case):
-        format = case["schema"].get("format")
-        return format not in checker.checkers or (
-            # datetime.datetime is overzealous about typechecking in <=1.9
-            format == "date-time" and
-            pypy_version_info is not None and
-            pypy_version_info[:2] <= (1, 9)
-        )
-    return missing_format
-
-
-class FormatMixin(object):
-    def test_it_returns_true_for_formats_it_does_not_know_about(self):
-        validator = self.validator_class(
-            {"format" : "carrot"}, format_checker=FormatChecker(),
-        )
-        validator.validate("bugs")
-
-    def test_it_does_not_validate_formats_by_default(self):
-        validator = self.validator_class({})
-        self.assertIsNone(validator.format_checker)
-
-    def test_it_validates_formats_if_a_checker_is_provided(self):
-        checker = mock.Mock(spec=FormatChecker)
-        validator = self.validator_class(
-            {"format" : "foo"}, format_checker=checker,
-        )
-
-        validator.validate("bar")
-
-        checker.check.assert_called_once_with("bar", "foo")
-
-        cause = ValueError()
-        checker.check.side_effect = FormatError('aoeu', cause=cause)
-
-        with self.assertRaises(ValidationError) as cm:
-            validator.validate("bar")
-        # Make sure original cause is attached
-        self.assertIs(cm.exception.cause, cause)
-
-
-@load_json_cases(
-    "draft3/*.json", ignore_glob=os.path.join("draft3", "refRemote.json")
-)
-@load_json_cases(
-    "draft3/optional/format.json", skip=missing_format(draft3_format_checker)
-)
-@load_json_cases("draft3/optional/bignum.json")
-@load_json_cases("draft3/optional/zeroTerminatedFloats.json")
-class TestDraft3(
-    unittest.TestCase, TypesMixin, DecimalMixin, FormatMixin
-):
-
-    validator_class = Draft3Validator
-    validator_kwargs = {"format_checker" : draft3_format_checker}
-
-    def test_any_type_is_valid_for_type_any(self):
-        validator = self.validator_class({"type" : "any"})
-        validator.validate(mock.Mock())
-
-    # TODO: we're in need of more meta schema tests
-    def test_invalid_properties(self):
-        with self.assertRaises(SchemaError):
-            validate({}, {"properties": {"test": True}},
-                     cls=self.validator_class)
-
-    def test_minItems_invalid_string(self):
-        with self.assertRaises(SchemaError):
-            # needs to be an integer
-            validate([1], {"minItems" : "1"}, cls=self.validator_class)
-
-
-@load_json_cases(
-    "draft4/*.json", ignore_glob=os.path.join("draft4", "refRemote.json")
-)
-@load_json_cases(
-    "draft4/optional/format.json", skip=missing_format(draft4_format_checker)
-)
-@load_json_cases("draft4/optional/bignum.json")
-@load_json_cases("draft4/optional/zeroTerminatedFloats.json")
-class TestDraft4(
-    unittest.TestCase, TypesMixin, DecimalMixin, FormatMixin
-):
-    validator_class = Draft4Validator
-    validator_kwargs = {"format_checker" : draft4_format_checker}
-
-    # TODO: we're in need of more meta schema tests
-    def test_invalid_properties(self):
-        with self.assertRaises(SchemaError):
-            validate({}, {"properties": {"test": True}},
-                     cls=self.validator_class)
-
-    def test_minItems_invalid_string(self):
-        with self.assertRaises(SchemaError):
-            # needs to be an integer
-            validate([1], {"minItems" : "1"}, cls=self.validator_class)
-
-
-class RemoteRefResolution(unittest.TestCase):
+class TestCreateAndExtend(unittest.TestCase):
     def setUp(self):
-        patch = mock.patch("jsonschema.requests")
-        requests = patch.start()
-        requests.get.side_effect = self.resolve
-        self.addCleanup(patch.stop)
+        self.meta_schema = {"properties" : {"smelly" : {}}}
+        self.smelly = mock.MagicMock()
+        self.validators = {"smelly" : self.smelly}
+        self.types = {"dict" : dict}
+        self.Validator = create(
+            meta_schema=self.meta_schema,
+            validators=self.validators,
+            default_types=self.types,
+        )
 
-    def resolve(self, reference):
-        _, _, reference = reference.partition("http://localhost:1234/")
-        return mock.Mock(**{"json.return_value" : REMOTES.get(reference)})
+        self.validator_value = 12
+        self.schema = {"smelly" : self.validator_value}
+        self.validator = self.Validator(self.schema)
 
+    def test_attrs(self):
+        self.assertEqual(self.Validator.VALIDATORS, self.validators)
+        self.assertEqual(self.Validator.META_SCHEMA, self.meta_schema)
+        self.assertEqual(self.Validator.DEFAULT_TYPES, self.types)
 
-@load_json_cases("draft3/refRemote.json")
-class Draft3RemoteResolution(RemoteRefResolution):
-    validator_class = Draft3Validator
+    def test_init(self):
+        self.assertEqual(self.validator.schema, self.schema)
 
+    def test_iter_errors(self):
+        instance = "hello"
 
-@load_json_cases("draft4/refRemote.json")
-class Draft4RemoteResolution(RemoteRefResolution):
-    validator_class = Draft4Validator
+        self.smelly.return_value = []
+        self.assertEqual(list(self.validator.iter_errors(instance)), [])
+
+        error = mock.Mock()
+        self.smelly.return_value = [error]
+        self.assertEqual(list(self.validator.iter_errors(instance)), [error])
+
+        self.smelly.assert_called_with(
+            self.validator, self.validator_value, instance, self.schema,
+        )
+
+    def test_if_a_version_is_provided_it_is_registered(self):
+        with mock.patch("jsonschema.validators.validates") as validates:
+            validates.side_effect = lambda version : lambda cls : cls
+            Validator = create(meta_schema={"id" : "id"}, version="my version")
+        validates.assert_called_once_with("my version")
+        self.assertEqual(Validator.__name__, "MyVersionValidator")
+
+    def test_if_a_version_is_not_provided_it_is_not_registered(self):
+        with mock.patch("jsonschema.validators.validates") as validates:
+            create(meta_schema={"id" : "id"})
+        self.assertFalse(validates.called)
+
+    def test_extend(self):
+        validators = dict(self.Validator.VALIDATORS)
+        new = mock.Mock()
+
+        Extended = extend(self.Validator, validators={"a new one" : new})
+
+        validators.update([("a new one", new)])
+        self.assertEqual(Extended.VALIDATORS, validators)
+        self.assertNotIn("a new one", self.Validator.VALIDATORS)
+
+        self.assertEqual(Extended.META_SCHEMA, self.Validator.META_SCHEMA)
+        self.assertEqual(Extended.DEFAULT_TYPES, self.Validator.DEFAULT_TYPES)
 
 
 class TestIterErrors(unittest.TestCase):
@@ -255,7 +89,7 @@ class TestIterErrors(unittest.TestCase):
             "minItems" : 3
         }
 
-        got = (str(e) for e in self.validator.iter_errors(instance, schema))
+        got = (e.message for e in self.validator.iter_errors(instance, schema))
         expected = [
             "%r is disallowed for [1, 2]" % (schema["disallow"],),
             "[1, 2] is too short",
@@ -353,12 +187,189 @@ class TestValidationErrorMessages(unittest.TestCase):
         self.assertIn("is not a", message)
 
 
-class TestValidationErrorDetails(unittest.TestCase):
-    def setUp(self):
-        self.validator = Draft3Validator({})
+class TestErrorReprStr(unittest.TestCase):
 
+    message = "hello"
+
+    def setUp(self):
+        self.error = ValidationError(
+            message=self.message,
+            validator="type",
+            validator_value="string",
+            instance=5,
+            schema={"type" : "string"},
+        )
+
+    def assertShows(self, message):
+        if PY3:
+            message = message.replace("u'", "'")
+        message = textwrap.dedent(message).rstrip("\n")
+
+        message_line, _, rest = str(self.error).partition("\n")
+        self.assertEqual(message_line, self.message)
+        self.assertEqual(rest, message)
+
+    def test_repr(self):
+        self.assertEqual(
+            repr(self.error),
+            "<ValidationError: %r>" % self.message,
+        )
+
+    def test_unset_error(self):
+        error = ValidationError("message")
+        self.assertEqual(str(error), "message")
+
+        kwargs = {
+            "validator": "type",
+            "validator_value": "string",
+            "instance": 5,
+            "schema": {"type": "string"}
+        }
+        # Just the message should show if any of the attributes are unset
+        for attr in kwargs:
+            k = dict(kwargs)
+            del k[attr]
+            error = ValidationError("message", **k)
+            self.assertEqual(str(error), "message")
+
+    def test_empty_paths(self):
+        self.error.path = self.error.schema_path = []
+        self.assertShows(
+            """
+            Failed validating u'type' in schema:
+                {u'type': u'string'}
+
+            On instance:
+                5
+            """
+        )
+
+    def test_one_item_paths(self):
+        self.error.path = [0]
+        self.error.schema_path = ["items"]
+        self.assertShows(
+            """
+            Failed validating u'type' in schema:
+                {u'type': u'string'}
+
+            On instance[0]:
+                5
+            """
+        )
+
+    def test_multiple_item_paths(self):
+        self.error.path = [0, "a"]
+        self.error.schema_path = ["items", 0, 1]
+        self.assertShows(
+            """
+            Failed validating u'type' in schema[u'items'][0]:
+                {u'type': u'string'}
+
+            On instance[0][u'a']:
+                5
+            """
+        )
+
+    def test_uses_pprint(self):
+        with mock.patch.object(pprint, "pformat") as pformat:
+            str(self.error)
+            self.assertGreater(pformat.call_count, 1)  # schema + instance
+
+
+class TestValidationErrorDetails(unittest.TestCase):
     # TODO: These really need unit tests for each individual validator, rather
     #       than just these higher level tests.
+    def test_anyOf(self):
+        instance = 5
+        schema = {
+            "anyOf": [
+                {"minimum": 20},
+                {"type": "string"}
+            ]
+        }
+
+        validator = Draft4Validator(schema)
+        errors = list(validator.iter_errors(instance))
+        self.assertEqual(len(errors), 1)
+        e = errors[0]
+
+        self.assertEqual(e.validator, "anyOf")
+        self.assertEqual(list(e.schema_path), ["anyOf"])
+        self.assertEqual(e.validator_value, schema["anyOf"])
+        self.assertEqual(e.instance, instance)
+        self.assertEqual(e.schema, schema)
+        self.assertEqual(list(e.path), [])
+        self.assertEqual(len(e.context), 2)
+
+        e1, e2 = sorted_errors(e.context)
+
+        self.assertEqual(e1.validator, "minimum")
+        self.assertEqual(list(e1.schema_path), [0, "minimum"])
+        self.assertEqual(e1.validator_value, schema["anyOf"][0]["minimum"])
+        self.assertEqual(e1.instance, instance)
+        self.assertEqual(e1.schema, schema["anyOf"][0])
+        self.assertEqual(list(e1.path), [])
+        self.assertEqual(len(e1.context), 0)
+
+        self.assertEqual(e2.validator, "type")
+        self.assertEqual(list(e2.schema_path), [1, "type"])
+        self.assertEqual(e2.validator_value, schema["anyOf"][1]["type"])
+        self.assertEqual(e2.instance, instance)
+        self.assertEqual(e2.schema, schema["anyOf"][1])
+        self.assertEqual(list(e2.path), [])
+        self.assertEqual(len(e2.context), 0)
+
+    def test_type(self):
+        instance = {"foo": 1}
+        schema = {
+            "type": [
+                {"type": "integer"},
+                {
+                    "type": "object",
+                    "properties": {
+                        "foo": {"enum": [2]}
+                    }
+                }
+            ]
+        }
+
+        validator = Draft3Validator(schema)
+        errors = list(validator.iter_errors(instance))
+        self.assertEqual(len(errors), 1)
+        e = errors[0]
+
+        self.assertEqual(e.validator, "type")
+        self.assertEqual(list(e.schema_path), ["type"])
+        self.assertEqual(e.validator_value, schema["type"])
+        self.assertEqual(e.instance, instance)
+        self.assertEqual(e.schema, schema)
+        self.assertEqual(list(e.path), [])
+        self.assertEqual(len(e.context), 2)
+
+        e1, e2 = sorted_errors(e.context)
+
+        self.assertEqual(e1.validator, "type")
+        self.assertEqual(list(e1.schema_path), [0, "type"])
+        self.assertEqual(e1.validator_value, schema["type"][0]["type"])
+        self.assertEqual(e1.instance, instance)
+        self.assertEqual(e1.schema, schema["type"][0])
+        self.assertEqual(list(e1.path), [])
+        self.assertEqual(len(e1.context), 0)
+
+        self.assertEqual(e2.validator, "enum")
+        self.assertEqual(
+            list(e2.schema_path),
+            [1, "properties", "foo", "enum"]
+        )
+        self.assertEqual(
+            e2.validator_value,
+            schema["type"][1]["properties"]["foo"]["enum"]
+        )
+        self.assertEqual(e2.instance, instance["foo"])
+        self.assertEqual(e2.schema, schema["type"][1]["properties"]["foo"])
+        self.assertEqual(list(e2.path), ["foo"])
+        self.assertEqual(len(e2.context), 0)
+
     def test_single_nesting(self):
         instance = {"foo" : 2, "bar" : [1], "baz" : 15, "quux" : "spam"}
         schema = {
@@ -369,7 +380,8 @@ class TestValidationErrorDetails(unittest.TestCase):
             }
         }
 
-        errors = self.validator.iter_errors(instance, schema)
+        validator = Draft3Validator(schema)
+        errors = validator.iter_errors(instance)
         e1, e2, e3, e4 = sorted_errors(errors)
 
         self.assertEqual(list(e1.path), ["bar"])
@@ -401,7 +413,8 @@ class TestValidationErrorDetails(unittest.TestCase):
             }
         }
 
-        errors = self.validator.iter_errors(instance, schema)
+        validator = Draft3Validator(schema)
+        errors = validator.iter_errors(instance)
         e1, e2, e3, e4, e5, e6 = sorted_errors(errors)
 
         self.assertEqual(list(e1.path), [])
@@ -424,7 +437,8 @@ class TestValidationErrorDetails(unittest.TestCase):
             "additionalProperties" : {"type": "integer", "minimum": 5}
         }
 
-        errors = self.validator.iter_errors(instance, schema)
+        validator = Draft3Validator(schema)
+        errors = validator.iter_errors(instance)
         e1, e2 = sorted_errors(errors)
 
         self.assertEqual(list(e1.path), ["bar"])
@@ -442,7 +456,8 @@ class TestValidationErrorDetails(unittest.TestCase):
             }
         }
 
-        errors = self.validator.iter_errors(instance, schema)
+        validator = Draft3Validator(schema)
+        errors = validator.iter_errors(instance)
         e1, e2 = sorted_errors(errors)
 
         self.assertEqual(list(e1.path), ["bar"])
@@ -458,7 +473,8 @@ class TestValidationErrorDetails(unittest.TestCase):
             "additionalItems" : {"type": "integer", "minimum": 5}
         }
 
-        errors = self.validator.iter_errors(instance, schema)
+        validator = Draft3Validator(schema)
+        errors = validator.iter_errors(instance)
         e1, e2 = sorted_errors(errors)
 
         self.assertEqual(list(e1.path), [0])
@@ -508,6 +524,27 @@ class TestErrorTree(unittest.TestCase):
         )
         tree = ErrorTree([e1, e2])
         self.assertEqual(tree["bar"][0].errors, {"foo" : e1, "quux" : e2})
+
+    def test_it_does_not_contain_subtrees_that_are_not_in_the_instance(self):
+        error = ValidationError("a message", validator="foo", instance=[])
+        tree = ErrorTree([error])
+
+        with self.assertRaises(IndexError):
+            tree[0]
+
+    def test_if_its_in_the_tree_anyhow_it_does_not_raise_an_error(self):
+        """
+        If a validator is dumb (like :validator:`required` in draft 3) and
+        refers to a path that isn't in the instance, the tree still properly
+        returns a subtree for that path.
+
+        """
+
+        error = ValidationError(
+            "a message", validator="foo", instance={}, path=["foo"],
+        )
+        tree = ErrorTree([error])
+        self.assertIsInstance(tree["foo"], ErrorTree)
 
 
 class ValidatorTestMixin(object):
@@ -589,6 +626,33 @@ class TestDraft4Validator(ValidatorTestMixin, unittest.TestCase):
     validator_class = Draft4Validator
 
 
+class TestValidatorFor(unittest.TestCase):
+    def test_draft_3(self):
+        schema = {"$schema" : "http://json-schema.org/draft-03/schema"}
+        self.assertIs(validator_for(schema), Draft3Validator)
+
+        schema = {"$schema" : "http://json-schema.org/draft-03/schema#"}
+        self.assertIs(validator_for(schema), Draft3Validator)
+
+    def test_draft_4(self):
+        schema = {"$schema" : "http://json-schema.org/draft-04/schema"}
+        self.assertIs(validator_for(schema), Draft4Validator)
+
+        schema = {"$schema" : "http://json-schema.org/draft-04/schema#"}
+        self.assertIs(validator_for(schema), Draft4Validator)
+
+    def test_custom_validator(self):
+        Validator = create(meta_schema={"id" : "meta schema id"}, version="12")
+        schema = {"$schema" : "meta schema id"}
+        self.assertIs(validator_for(schema), Validator)
+
+    def test_validator_for_jsonschema_default(self):
+        self.assertIs(validator_for({}), Draft4Validator)
+
+    def test_validator_for_custom_default(self):
+        self.assertIs(validator_for({}, default=None), None)
+
+
 class TestValidate(unittest.TestCase):
     def test_draft3_validator_is_chosen(self):
         schema = {"$schema" : "http://json-schema.org/draft-03/schema#"}
@@ -614,10 +678,14 @@ class TestValidate(unittest.TestCase):
 
 
 class TestRefResolver(unittest.TestCase):
+
+    base_uri = ""
+    stored_uri = "foo://stored"
+    stored_schema = {"stored" : "schema"}
+
     def setUp(self):
-        self.base_uri = ""
         self.referrer = {}
-        self.store = {}
+        self.store = {self.stored_uri : self.stored_schema}
         self.resolver = RefResolver(self.base_uri, self.referrer, self.store)
 
     def test_it_does_not_retrieve_schema_urls_from_the_network(self):
@@ -633,7 +701,18 @@ class TestRefResolver(unittest.TestCase):
         with self.resolver.resolving(ref) as resolved:
             self.assertEqual(resolved, self.referrer["properties"]["foo"])
 
+    def test_it_resolves_local_refs_with_id(self):
+        schema = {"id": "foo://bar/schema#", "a": {"foo": "bar"}}
+        resolver = RefResolver.from_schema(schema)
+        with resolver.resolving("#/a") as resolved:
+            self.assertEqual(resolved, schema["a"])
+        with resolver.resolving("foo://bar/schema#/a") as resolved:
+            self.assertEqual(resolved, schema["a"])
+
     def test_it_retrieves_stored_refs(self):
+        with self.resolver.resolving(self.stored_uri) as resolved:
+            self.assertIs(resolved, self.stored_schema)
+
         self.resolver.store["cached_ref"] = {"foo" : 12}
         with self.resolver.resolving("cached_ref#/foo") as resolved:
             self.assertEqual(resolved, 12)
@@ -642,7 +721,7 @@ class TestRefResolver(unittest.TestCase):
         ref = "http://bar#baz"
         schema = {"baz" : 12}
 
-        with mock.patch("jsonschema.requests") as requests:
+        with mock.patch("jsonschema.validators.requests") as requests:
             requests.get.return_value.json.return_value = schema
             with self.resolver.resolving(ref) as resolved:
                 self.assertEqual(resolved, 12)
@@ -652,8 +731,8 @@ class TestRefResolver(unittest.TestCase):
         ref = "http://bar#baz"
         schema = {"baz" : 12}
 
-        with mock.patch("jsonschema.requests", None):
-            with mock.patch("jsonschema.urlopen") as urlopen:
+        with mock.patch("jsonschema.validators.requests", None):
+            with mock.patch("jsonschema.validators.urlopen") as urlopen:
                 urlopen.return_value.read.return_value = (
                     json.dumps(schema).encode("utf8"))
                 with self.resolver.resolving(ref) as resolved:
@@ -664,13 +743,23 @@ class TestRefResolver(unittest.TestCase):
         schema = {"id" : "foo"}
         resolver = RefResolver.from_schema(schema)
         self.assertEqual(resolver.base_uri, "foo")
-        self.assertEqual(resolver.referrer, schema)
+        with resolver.resolving("") as resolved:
+            self.assertEqual(resolved, schema)
+        with resolver.resolving("#") as resolved:
+            self.assertEqual(resolved, schema)
+        with resolver.resolving("foo") as resolved:
+            self.assertEqual(resolved, schema)
+        with resolver.resolving("foo#") as resolved:
+            self.assertEqual(resolved, schema)
 
     def test_it_can_construct_a_base_uri_from_a_schema_without_id(self):
         schema = {}
         resolver = RefResolver.from_schema(schema)
         self.assertEqual(resolver.base_uri, "")
-        self.assertEqual(resolver.referrer, schema)
+        with resolver.resolving("") as resolved:
+            self.assertEqual(resolved, schema)
+        with resolver.resolving("#") as resolved:
+            self.assertEqual(resolved, schema)
 
     def test_custom_uri_scheme_handlers(self):
         schema = {"foo": "bar"}
@@ -684,8 +773,9 @@ class TestRefResolver(unittest.TestCase):
     def test_cache_remote_on(self):
         ref = "foo://bar"
         foo_handler = mock.Mock()
-        resolver = RefResolver("", {}, cache_remote=True,
-                               handlers={"foo": foo_handler})
+        resolver = RefResolver(
+            "", {}, cache_remote=True, handlers={"foo" : foo_handler},
+        )
         with resolver.resolving(ref):
             pass
         with resolver.resolving(ref):
@@ -695,58 +785,29 @@ class TestRefResolver(unittest.TestCase):
     def test_cache_remote_off(self):
         ref = "foo://bar"
         foo_handler = mock.Mock()
-        resolver = RefResolver("", {}, cache_remote=False,
-                               handlers={"foo": foo_handler})
+        resolver = RefResolver(
+            "", {}, cache_remote=False, handlers={"foo" : foo_handler},
+        )
         with resolver.resolving(ref):
             pass
         with resolver.resolving(ref):
             pass
         self.assertEqual(foo_handler.call_count, 2)
 
-
-class TestFormatChecker(unittest.TestCase):
-    def setUp(self):
-        self.fn = mock.Mock()
-
-    def test_it_can_validate_no_formats(self):
-        checker = FormatChecker(formats=())
-        self.assertFalse(checker.checkers)
-
-    def test_it_raises_a_key_error_for_unknown_formats(self):
-        with self.assertRaises(KeyError):
-            FormatChecker(formats=["o noes"])
-
-    def test_it_can_register_cls_checkers(self):
-        with mock.patch.dict(FormatChecker.checkers, clear=True):
-            FormatChecker.cls_checks("new")(self.fn)
-            self.assertEqual(FormatChecker.checkers, {"new" : (self.fn, ())})
-
-    def test_it_can_register_checkers(self):
-        checker = FormatChecker()
-        checker.checks("new")(self.fn)
-        self.assertEqual(
-            checker.checkers,
-            dict(FormatChecker.checkers, new=(self.fn, ()))
-        )
-
-    def test_it_catches_registered_errors(self):
-        checker = FormatChecker()
-        checker.checks("foo", raises=ValueError)(self.fn)
-        # Registered errors should be caught and turned into FormatErrors
-        cause = ValueError()
-        self.fn.side_effect = cause
-        with self.assertRaises(FormatError) as cm:
-            checker.check("bar", "foo")
-        # Original exception should be attached to cause attribute
-        self.assertIs(cm.exception.cause, cause)
-        # Unregistered errors should not be caught
-        self.fn.side_effect = AttributeError
-        with self.assertRaises(AttributeError):
-            checker.check("bar", "foo")
-
+    def test_if_you_give_it_junk_you_get_a_resolution_error(self):
+        ref = "foo://bar"
+        foo_handler = mock.Mock(side_effect=ValueError("Oh no! What's this?"))
+        resolver = RefResolver("", {}, handlers={"foo" : foo_handler})
+        with self.assertRaises(RefResolutionError) as err:
+            with resolver.resolving(ref):
+                pass
+        self.assertEqual(str(err.exception), "Oh no! What's this?")
 
 
 def sorted_errors(errors):
     def key(error):
-        return [str(e) for e in error.path], error.validator
+        return (
+            [str(e) for e in error.path],
+            [str(e) for e in error.schema_path]
+        )
     return sorted(errors, key=key)